Comment rating optimization

+++ ms-comment-rating/src/main/java/cl/govegan/ms_comment_rating/services/impl/RatingServiceImpl.java
@@ -1,10 +1,10 @@
 package cl.govegan.ms_comment_rating.services.impl;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import cl.govegan.ms_comment_rating.exceptions.ResourceNotFoundException;
 import cl.govegan.ms_comment_rating.models.Rating;
 import cl.govegan.ms_comment_rating.repositories.RatingRepository;
 import cl.govegan.ms_comment_rating.services.RatingServices;
@@ -14,13 +14,13 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class RatingServiceImpl implements RatingServices{
 
-   @Autowired
-   private RatingRepository ratingRepository;
+   private final RatingRepository ratingRepository;
 
    @Override
-   public Rating findById(String id) {
-      return ratingRepository.findById(id).orElse(null);
-   }
+    public Rating findById(String id) {
+        return ratingRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Rating not found with id: " + id));
+    }
 
    @Override
     public Page<Rating> findByRecipeId(String recipeId, Pageable pageable) {
@@ -37,29 +37,34 @@ public class RatingServiceImpl implements RatingServices{
         return ratingRepository.findByUsernameAndRecipeId(username, recipeId, pageable);
     }
 
-   @Override
-   public Rating addRating(Rating rating) {
-      return ratingRepository.save(rating);
-   }
+    @Override
+    public Rating addRating(Rating rating) {
+        return ratingRepository.save(rating);
+    }
 
-   @Override
-   public Rating updateRating(Rating rating) {
-      return ratingRepository.save(rating);
-   }
-
-   @Override
-   public void deleteRating(String id) {
-      ratingRepository.deleteById(id);
-   }
-   
-   @Override
-    public double getAverageRatingByRecipeId(String recipeId) {
-        Page<Rating> ratings = ratingRepository.findByRecipeId(recipeId, Pageable.ofSize(10));
-        double sum = 0;
-        for (Rating rating : ratings.getContent()) {
-            sum += rating.getRating();
+    @Override
+    public Rating updateRating(Rating rating) {
+        if (!ratingRepository.existsById(rating.getId())) {
+            throw new ResourceNotFoundException("Rating not found");
         }
-        return sum / ratings.getTotalElements();
+        return ratingRepository.save(rating);
+    }
+
+    @Override
+    public void deleteRating(String id) {
+        if (!ratingRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Rating not found");
+        }
+        ratingRepository.deleteById(id);
+    }
+   
+    @Override
+    public double getAverageRatingByRecipeId(String recipeId) {
+        Double averageRating = ratingRepository.findAverageRatingByRecipeId(recipeId);
+        if (averageRating == null) {
+            throw new ResourceNotFoundException("No ratings found for recipeId: " + recipeId);
+        }
+        return averageRating;
     }
     
 }
